Added an (optional) debugging version of dbref selectable via a #define and made the code compatible with both the normal dbref and the debugging one Made Compound_command_and_arguments intialise the 'player' dbref.

+++ alarm.c
@@ -44,7 +44,7 @@ db_patch_alarms ()
 		{
 			context *c = new context (GOD_ID);
 			char tmpnum[20];
-			sprintf(tmpnum, "#%d", i);
+			sprintf(tmpnum, "#%d", (int)i);
 			db[i].clear_flag(FLAG_CONNECTED);
 			c->do_at_connect(tmpnum, NULLSTRING);
 			delete mud_scheduler.push_express_job (c);
