+++ sanity-check.c
@@ -73,12 +73,12 @@ sane_unparse_object (dbref thing)
 
 	if ((thing < 0 || thing >= db.get_top()) && thing != NOTHING && thing != HOME)
 	{
-		sprintf (buf, "***ILLEGAL (%d)***", thing);
+		sprintf (buf, "***ILLEGAL (%d)***", (int)thing);
 		return (buf);
 	}
 	else if(Typeof(thing) == TYPE_FREE)
 	{
-		sprintf (buf, "***FREE (%d)***", thing);
+		sprintf (buf, "***FREE (%d)***", (int)thing);
 		return (buf);
 	}
 	else
@@ -125,14 +125,14 @@ int		fields)
 		if (Typeof (i) == TYPE_PLAYER)
 			fprintf(sanity_file, "\tController: %s\n", sane_unparse_object (db[i].get_controller()));
 		else
-			fprintf(sanity_file, "\tController: %d is not a player, so no controller field!\n", i);
+			fprintf(sanity_file, "\tController: %d is not a player, so no controller field!\n", (int)i);
 	}
 /*	if (fields & FLAGS)
 		fprintf(sanity_file, "\tFlags: %x\n", db[i].get_flags());*/
 	if (fields & TYPE)
 		fprintf(sanity_file, "\tType: %x\n", Typeof (i));
 	if (fields & PARENT)
-		fprintf(sanity_file, "\tParent: %s(%d)\n", getname (db[i].get_parent ()), db[i].get_parent ());
+		fprintf(sanity_file, "\tParent: %s(%d)\n", getname (db[i].get_parent ()), (int)(db[i].get_parent ()));
 }
 
 
@@ -153,7 +153,7 @@ dbref	owner)
 	{
 		if (In_a_contents_list (temp_list) && (sanity_array [temp_list].first_list != owner))
 		{
-			sprintf (buffer, "CONTENTS RULES: IN LISTS FOR #%d AND #%d", sanity_array [temp_list].first_list, owner);
+			sprintf (buffer, "CONTENTS RULES: IN LISTS FOR #%d AND #%d", (int)(sanity_array [temp_list].first_list), (int)owner);
 			violate (prev_list, buffer, OWNER|LOCATION);
 			fatal++;
 		}
@@ -190,7 +190,7 @@ dbref	owner)
 			sanity_array [temp_list].flags &= ~(SANITY_MEMBER_HAS_FOUND);
 			if (db [temp_list].get_location() == NOTHING)
 			{
-				sprintf (buffer, "POINTER RULES: OBJECT HAS NO LOCATION, BUT IS IN LIST for %d", sanity_array [temp_list].first_list);
+				sprintf (buffer, "POINTER RULES: OBJECT HAS NO LOCATION, BUT IS IN LIST for %d", (int)(sanity_array [temp_list].first_list));
 				violate (temp_list, buffer, 0);
 				fatal++;
 			}
